feat: Não deve ser possível editar ou excluir veículos que estejam com aluguel em curso feat: obter todos os veículos disponíveis para locação no momento atual

+++ src/main/java/com/br/luggycar/api/services/VehicleService.java
@@ -1,11 +1,17 @@
 package com.br.luggycar.api.services;
 
+import com.br.luggycar.api.dtos.response.ClientResponse;
 import com.br.luggycar.api.dtos.response.VehicleResponse;
 import com.br.luggycar.api.entities.Category;
 import com.br.luggycar.api.entities.Client;
 import com.br.luggycar.api.entities.Vehicle;
+import com.br.luggycar.api.enums.rent.RentStatus;
+import com.br.luggycar.api.exceptions.ResourceDatabaseException;
+import com.br.luggycar.api.exceptions.ResourceExistsException;
 import com.br.luggycar.api.exceptions.ResourceNotFoundException;
 import com.br.luggycar.api.repositories.CategoryRepository;
+import com.br.luggycar.api.repositories.ClientRepository;
+import com.br.luggycar.api.repositories.RentRepository;
 import com.br.luggycar.api.repositories.VehicleRepository;
 import com.br.luggycar.api.dtos.requests.VehicleRequest;
 import org.springframework.beans.BeanUtils;
@@ -24,11 +30,10 @@ public class VehicleService {
     private VehicleRepository vehicleRepository;
     @Autowired
     private CategoryRepository categoryRepository;
+    @Autowired
+    private RentRepository rentRepository;
 
     public VehicleResponse createVehicle(VehicleRequest vehicleRequest) {
-
-//        Category category = categoryRepository.findById(vehicleRequest.categoryId())
-//                .orElseThrow(() -> new RuntimeException("Categoria não encontrada"));
 
         Category category = categoryRepository.findByName(vehicleRequest.categoryName())
                 .orElseThrow(() -> new RuntimeException("Categoria não encontrada"));
@@ -57,6 +62,12 @@ public class VehicleService {
         Vehicle vehicle = vehicleRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Veículo não encontrado"));
 
+        List<RentStatus> activeStatuses = List.of(RentStatus.PENDING, RentStatus.IN_PROGRESS);
+
+        if (rentRepository.existsByVehicleIdAndStatusIn(id,activeStatuses)) {
+            throw new ResourceExistsException("veículo com aluguel em curso não pode ser editado.");
+        }
+
         BeanUtils.copyProperties(vehicleRequest, vehicle, "id", "registrationDate");
 
         Vehicle updatedVehicle = vehicleRepository.save(vehicle);
@@ -65,14 +76,36 @@ public class VehicleService {
     }
 
     public void deleteVehicle(Long id) {
+
+        List<RentStatus> activeStatuses = List.of(RentStatus.PENDING, RentStatus.IN_PROGRESS);
+
+        if (rentRepository.existsByVehicleIdAndStatusIn(id,activeStatuses)) {
+            throw new ResourceExistsException("veículo com aluguel em curso não pode ser excluido.");
+        }
+
         vehicleRepository.deleteById(id);
     }
-
 
     public Optional<VehicleResponse> findVehicleById(Long id) {
         return vehicleRepository.findById(id)
                 .map(VehicleResponse::new);
     }
 
+    public List <VehicleResponse>getAvailableVehicles() {
 
+        try {
+            List<Vehicle> allVehicles = vehicleRepository.findAll();
+            List<RentStatus> activeStatuses = List.of(RentStatus.PENDING, RentStatus.IN_PROGRESS);
+            List<Vehicle> rentedVehicles = rentRepository.findRentedVehiclesByStatusIn(activeStatuses);
+            allVehicles.removeAll(rentedVehicles);
+
+            return allVehicles.stream()
+                    .map(VehicleResponse::new)
+                    .collect(Collectors.toList());
+
+        } catch (
+                ResourceDatabaseException e) {
+            throw new ResourceDatabaseException("Erro ao buscar os veiculos diponiveis para aluguel no banco de dados", e);
+        }
+    }
 }
